requirements.txt
discord.py
supabase
flask
pynacl
matplotlib
gunicorn